requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
--index-url https://pypi.python.org/simple

cached-property==1.5.2
    # via h5py
docutils==0.16
    # via -r requirements.in
h5py==3.1.0
    # via -r requirements.in
joblib==1.0.1
    # via scikit-learn
nmslib==2.1.1
    # via -r requirements.in
numpy==1.20.1
    # via
    #   -r requirements.in
    #   h5py
    #   nmslib
    #   pandas
    #   pytorch-metric-learning
    #   scikit-learn
    #   scipy
    #   torch
    #   torchvision
pandas==1.2.2
    # via -r requirements.in
pillow==8.1.0
    # via torchvision
psutil==5.8.0
    # via nmslib
pybind11==2.6.1
    # via nmslib
python-dateutil==2.8.1
    # via pandas
pytorch-metric-learning==0.9.96
    # via -r requirements.in
pytz==2021.1
    # via pandas
scikit-learn==0.24.1
    # via pytorch-metric-learning
scipy==1.6.1
    # via
    #   -r requirements.in
    #   scikit-learn
six==1.15.0
    # via python-dateutil
threadpoolctl==2.1.0
    # via scikit-learn
torch==1.7.1
    # via
    #   -r requirements.in
    #   pytorch-metric-learning
    #   torchvision
torchvision==0.8.2
    # via
    #   -r requirements.in
    #   pytorch-metric-learning
tqdm==4.57.0
    # via pytorch-metric-learning
typing-extensions==3.7.4.3
    # via torch
